// src/types/generated/ScheduleLibraryItem.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ScheduleCommandDto } from "./ScheduleCommandDto";

/**
 * A schedule library item (template with embedded commands)
 */
export type ScheduleLibraryItem = { id: number, site_id: number, name: string, description: string | null, commands: Array<ScheduleCommandDto>, created_at: string, };